notmuch: remove scroll bindings
After thinking about it some more I'm a bit reluctant about the zz bindings. My
"real" goal was to have normal-mode-like feel when reading email. Such as
jumping around paragraphs with { and }, etc. I'm not sure if there's a way to
retain normal-mode bindings somehow.

I'm not that familiar with evilified states so I'm not even sure if this is
possible. For now I think we shouldn't re-add normal mode mappings to the
notmuch-show evilified state as that seems a bit ad-hoc...

@braham-snyder mentioned[1] that was working on something along these lines

[1] https://github.com/syl20bnr/spacemacs/issues/2163#issuecomment-325036013